move some docker methods from bootstrap and commons to docker client, move run functions to separate module
