+++ tests/unit/test_misc.py
@@ -12,8 +12,8 @@ from localstack.services import infra
 from localstack.services.generic_proxy import GenericProxy, ProxyListener
 from localstack.utils import async_utils, config_listener
 from localstack.utils.aws import aws_stack
-from localstack.utils.bootstrap import PortMappings
 from localstack.utils.common import TMP_FILES, download, json_safe, load_file, now_utc, parallelize
+from localstack.utils.docker import PortMappings
 from localstack.utils.http_utils import create_chunked_data, parse_chunked_data
